RelativeDateTimeFormat (rdtf) should produce the same results as ICU4J's RelativeDateTimeFormatter in all cases. http://icu-project.org/apiref/icu4j/com/ibm/icu/text/RelativeDateTimeFormatter.html
However, ICU4J has special cases for 0 in the units MINUTE and HOUR,
as described in the source code.

This CL fixes Closure's rdtf to be consistent with ICU4J's output.

RELNOTES: Some user visible change, but only for MINUTE and HOUR when the quantity is 0, i.e., "this minute" changes to "in 1 minute" and "this hour" changes to "in 1 hour".

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=239220648