Replace calls to goog.define with assignments to the result of goog.define.
Background: To support goog.define within modules, it will soon stop exporting the named variable on the global object. In preparation for this, we need to manually export the names instead. This is effectively performing the rewriting that the compiler was already doing in RewriteClosurePrimitives.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=239512874