Ensure the closure_deps/ package's compiler and library dependencies are up to date.
Note: for the library the dep was removed entirely and instead tests will be run assuming that the test is running from the git repo. We have to do this because we now have an .npmignore file in the library which ignores some files that end up in the deps.js file. So if we run against the npm version we can't generate a correct deps.js file due to the missing sources.

TODO to remove the library dep from the package.json.

RELNOTES: n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=239682296