Adding SafeUrl.isSafeMimeType().
This enables a client to determine in advance whether a Blob is obviously unsafe based on its type. This is useful for Bulletin so we can reject weird file types (like .avi or .jp2) before attempting to display them to the user and encountering errors from SafeUrl.

RELNOTES[NEW]: Adding SafeUrl.isSafeMimeType() to determine whether or not a MIME type is considered safe before attempting to construct a SafeUrl. This is helpful for users who want to gracefully filter out bad MIME types before committing to creating a SafeUrl.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=243085956