Make goog.base safe for ES5 classes that have ES6 subclasses.
The is issue is that superClass_ is a static property on the ES5 class, so that gets statically inherited on the ES6 subclasses. Meaning you can end up with this:

ES5 class A
ES5 class B extends A (superClass_ is A)
ES6 class C extends B (superClass_ is **A**)
... other ES6 [transitive] subclasses of C also have a superClass_ of A.

So if C calls a super method in B (with ES6 `super`), which then uses goog.base to call a super method in A, it will search for the "current caller" first on C, then its superClass_, which is A. As a result it fails to find the caller, B.

Note: It is still invalid to call goog.base within the ES6 class itself. Use `super` instead.

RELNOTES[INC]: Fix goog.base not working in ES5 classes that have ES6 subclasses. This is not IE8 compatible and other pre-ES5 javascript environments.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=245128402