RELNOTES: Fix an IE11 failure in Closure's sanitizer when templates are polyfilled
This line of code interacts badly with the Polymer polyfill: the sanitizer assumes templates are available while the ersatz is not entirely the same, and that fails later on with cryptic errors. To mitigate, gate this branch on the browser instead of detecting features: that way, polyfills shouldn't affect the code.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=249013640