Add wait function that will allow you to get a promise that resolves when a given function is called.
This CL adds the a dictionary `callListeners` which is keyed by 'methodName' and whose values contain a list of callbacks. Anytime a function call is recorded we will invoke all of the corresponding call listeners.

The wait function will return a promise that will resolve when the callback it adds to the `callListeners` dictionary is called the correct number of times with the correct parameters.

RELNOTES[NEW]: goog.labs.mock now has a top level wait function. This function returns a promise that resolves when a provided function is called.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=251308041