Pipe response status through the BulkLoader and ModuleLoader error handlers and log it along with the module request error in Ritz.
Previously, the error status was not being passed by the BulkLoader to the error event, and the ModuleLoader expected an error status where it was actually receiving an error event object. As a result, the ModuleManager wasn't properly receiving the error status in this error handlers.

Now, that error status is piped through to the ModuleLoader properly, and it is added to the ModuleLoader's RequestError event. It may still be null if the error was not related to an HTTP error.

RELNOTES[NEW]: Fixed a bug in goog.module.ModuleLoader's request error handling, and added the response status to the goog.module.ModuleLoader.RequestErrorEvent.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=251855320