Enable Base 64 encodings to be given an alphabet
Instead of opt_webSafe, provides a flexible way to choose an alphabet with an enum.

RELNOTES: Enables Base 64 encodings to be given an alphabet instead of a boolean flag to specify if a websafe encoding is used. A new `Alphabet` enum is added to be used for the parameter. The legacy `opt_webSafe=true` is softly deprecated, but will still be translated into `alphabet=Alphabet.WEBSAFE_DOT_PADDING` for compatibility.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=258232816