Allowing alpha values of '1.0' (as opposed to only '1') in rgba (and hsla) Javascript string colors, e.g. 'rgba(255,0,0,1.0)' is now a valid color.
In rgba color values, alpha is a float a\in[0,1]. Writing the alpha value for no transparency as 1.0 instead of 1 feels the most intuitive to me (and our research/loco/vision codebase uses it extensively, which motivated this CL).
The current regex does not match a=1.0 though.

RELNOTES: Allowing alpha values of '1.0' (as opposed to only '1') in rgba (and hsla) Javascript string colors.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=259929345