Fix bug in Closure MenuButton menu positioning
When a menu button is near the right edge, and the browser window is zoomed in, decreasing the viewport size, the menu can get temporarily squashed by the browser, causing the code to use an invalid width to compute the new menu position. This CL inserts an initial step moving the menu to the top left of its container, letting it lay out properly, before continuing with the repositioning.

RELNOTES: Fixed a bug in the MenuButton class. When the browser magnification is increased (decreasing the viewport size), and a menu button close to the right edge has an open popup menu, the menu could be improperly repositioned in response to the zoom.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=265133812