Change PopupDatePicker position to AnchoredViewportPosition.
Thanks to this, if the date input is at the end of viewport and there is no place to show it below the input, the popup will be displayed above the input, not on top of it.

RELNOTES: goog.ui.PopupDatePicker now uses AnchoredViewportPosition instead of AnchoredPosition.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=269757882