Add goog.module to mochikit/async tests
This also removes deferredTest::testCallbackScope2, which tests to make sure that the value of 'this' inside a deferred callback is equal to window. As a module this is no longer true (except in IE9). Arguably the real behavior being tested (user can specify a scope) is already covered in testCallbackScope1 and testCallbackScope3. (see: cl/11714070)

RELNOTES: n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=269947509