Set 'aria-selected' to true for the caption of goog.ui.Select
The caption was assigned role="option" to improve verbalization (see CL 69770216). But if we don't set 'aria-selected' to true, some screen readers, e.g. ChromeVox, will append "not selected" to the verbalization. We don't want this since the current caption of the select is apparently the selected option.

RELNOTES: n/a

PiperOrigin-RevId: 282771527