Stop special-casing for IE in Element#getBoundingClientRect.
This was causing issues for some versions of IE and Edge, where assigning a field would throw an error. rect.* values are read-only, so this branch has effectively been a no-op even if an error wasn't thrown. In either case, the special case was introduced for IE 6/7 in 2008 and doesn't apply anymore.

RELNOTES: n/a
PiperOrigin-RevId: 310626058