Copy much of the interface of `Deferred` onto `Result`.
This allows uses of `Result` to be rewritten to look like `Deferred`. Doing so simplifies migration. Files can be updated one-by-one to call into the `Deferred`-like methods, without changing any type signatures. Then all the type annotations in a project can be updated together as a trivial change.

RELNOTES: Copy much of the interface of `Deferred` onto `Result`.
PiperOrigin-RevId: 313421922