Make goog.testing.jsunit.AUTO_RUN_DELAY_IN_MS explicitly writable and make a new define value goog.testing.jsunit.AUTO_RUN_DELAY_IN_MS_DEFAULT.
The value goog.testing.jsunit.AUTO_RUN_DELAY_IN_MS is overridden in the tests that require a longer settle time before the tests run and this is typically set in the test itself.  However, as a define value the compiler complains when the value is written in the test as they are expected to be constant. This makes a speed bump to having the tests be type checked.

RELNOTES: The goog.testing.jsunit.AUTO_RUN_DELAY_IN_MS define value is now goog.testing.jsunit.AUTO_RUN_DELAY_IN_MS_DEFAULT
PiperOrigin-RevId: 334240983