Fix incorrect results from adding DateTime intervals that span over DST start or end.
This is due to using the local date instead of UTC date to calculate the base date to which the interval should be added. Note that UTC time is already used instead of local time, so this should make things consistent across date and time as well.

RELNOTES: UTC dates are now used (instead of local dates) when adding intervals to goog.date.DateTime objects.
PiperOrigin-RevId: 335801398