Make the function parameter argument to goog.labs.mock.mockFunction optional
This aligns it with goog.testing.recordFunction and is pretty straightforward given that the function argument was only used for getting a string from its name.  A future change might possibly accept just a string, but it's not yet clear what the uniqueness requirements are (given that it uses getUid to uniquify anonymous function names).

RELNOTES: goog.labs.mock.mockFunction parameter is now optional.
PiperOrigin-RevId: 339922897