Deprecate goog.tweak in preparation for turning down the unused bits of it.
goog.tweak was meant to be an alternative to goog.define but never saw adoption. goog.tweak is more "framework-y" that Closure Library wants to be.  goog.define fulfills many of the use cases that goog.tweak might fulfill (the ability to set define value without needing to be first processed by the compiler) but without the UI.  A simple bootstrap script could easily be written to parse the query and set CLOSURE_DEFINES based on it.

RELNOTES: Mark goog.tweak as deprecated.
PiperOrigin-RevId: 346443998