Fix PerformanceTimer.runAsyncTask to not use recursive callbacks.
The recursive callback approach pollutes the stack and makes it difficult to
profile performance across iterations since the stack grows one level deeper on
each iteration. Instead, I've converted to code to use a deferred instance for
each executable step regardless of whether that step is async or not.
RELNOTES: Fix PerformanceTimer.runAsyncTask to not use recursive callbacks.
PiperOrigin-RevId: 357031582