Implement a predictive parsing option for goog.i18n.DateTimeParse with support for limited pattern symbols.
The `predictive` option provides better support for parsing live user inputs where parsing occurs as the user types, and it is therefore desirable to successfully parse those partial inputs to value for which the input is a prefix. E.g. for the pattern 'h:mm a' the input '9:4' will parse to a value equivalent to '9:40AM' instead of '9:04AM' and '9:40 p' will parse to '9:40PM' instead of '9:40AM'.

It may be desirable for the purpose described above to use the `predictive` option in combination with the `validate` option to enforce that inputs fall within the standard time of day range. E.g. causing parsing of '123:456' to fail.

RELNOTES[NEW]: Implement a predictive parsing option for goog.i18n.DateTimeParse with support for limited pattern symbols.

PiperOrigin-RevId: 366933947