make some type annotations explicit
This CL adds an explicit type to some code that lacked one.
The intent is to preserve the existing type, but make it explicit.
(Note that in the case of function return types, Closure sometimes
infers a `{?}` return type, so this CL may tighten some of these.)
RELNOTES: n/a
PiperOrigin-RevId: 367694623