Fix Date interval addition for years 0-99
The Closure implementation for adding an interval to a
Closure Date instance delegates to the Javascript Date
type.  The 'year' argument to the Javascript Date
constructor has special behavior for years 0-99:

 * those values are interpreted as being relative to
   the year 1900
 * all other values are interpreted as being absolute

This patch corrects for that specific behavior when
adding intervals to dates where a year falls within
the problematic range.

RELNOTES: Fix Date.add behavior for years 0-99.
PiperOrigin-RevId: 391791610
Change-Id: I5812b11d51602e035802b47852f136c61387cf18