Fix a daylight saving time bug in `formatDay`.
If the current date-time is 23:MM PST on the day daylight saving time ends, and the passed in date-time is also 23:MM PST on the day daylight saving time ends (possibly for some other MM), then `formatDay` returns 'Tomorrow' rather than 'Today'. See the explanation in the comment added in this CL for why (and for the explanation of the fix).
RELNOTES: Fix a daylight saving time bug in `formatDay`.
PiperOrigin-RevId: 417926514
Change-Id: I2eefc039f16584abcbf4b321349dd18918805995