Create @define goog.html.trustedtypes.POLICY_NAME
The purpose of this is to allow binaries to set a policy name for Safe Types that doesn't end with `#html`. This is most useful for SDKs intended to be used by third parties.

RELNOTES[NEW]: Add `goog.html.trustedtypes.POLICY_NAME` as a more specific alternative to `goog.TRUSTED_TYPES_POLICY_NAME`.

PiperOrigin-RevId: 427809795
Change-Id: I7860e3ae48068852e114fed3664c7ffd360f8f1d