Deprecate the `goog.getLocale()` method
This method was added to support a project that ended up not
using it after all.

Any code that started using it should switch back to `goog.LOCALE`.

We should remove it, because closure-compiler requires an extra
step (inlining the method) to optimize code that uses it.

RELNOTES: Deprecated `goog.getLocale()` - use `goog.LOCALE` instead
PiperOrigin-RevId: 434873631
Change-Id: I48b3357447d82145a447925a77f1224fe34f397b