Stop escaping periods in Closure release notes markdown generation
This has been problematic because, in practices, most periods are in qualified names, which are (or should be) in code blocks, where period-escaping doesn't apply (so we end up with extra backslashes).  Ideally we would have a smarter processor, but it's very difficult to infer what authors intended when writing release notes.

RELNOTES: n/a
PiperOrigin-RevId: 436370208
Change-Id: I1b4fa57b61df33389f7b9a9378dbf5195bdec1d3