Fix missing require issue in goog.a11y.area.datatables.
Fixes https://github.com/bazelbuild/rules_closure/issues/550, which seems to have been caused by a change last summer to remove the `missingRequire` suppression, even though the underlying issue wasn't actually fixed.  It seems that somewhere along the way something got moved, but somebody forgot to update the static variable to be in the provided namespace.

Rather than moving the static variable into the right namespace, I've made this a `goog.module` so that we can just use a module-local for the same purpose.

RELNOTES: n/a
PiperOrigin-RevId: 438380355
Change-Id: I104b4f500af9c275fdcef2512de7c11b882a02cd