Refactor `goog.crypt.stringToByteArray` and related methods
The problem with the existing methods is that the string encoding is ambiguous and currently lossy: characters beyond 255 are encoded little-endian with no indication of the multibyte character, making it impossible to reverse.

We pull out new methods with clearer semantics:

* `goog.crypt.textToByteArray`
* `goog.crypt.byteArrayToText`
* `goog.crypt.binaryStringToByteArray`
* `goog.crypt.byteArrayToBinaryString`
* `goog.base64.encodeText`
* `goog.base64.decodeToText`
* `goog.base64.encodeBinaryString`
* `goog.base64.decodeToBinaryString`
* `goog.hash32.encodeText`
* `goog.hash32.encodeBinaryString`

The existing methods that refer to simply "string" (including those that specifically include "utf8") will be deprecated. These methods have three different behaviors for multi-byte characters:

The "text" methods do a full UTF-8 encoding, which will change the result for characters in the [128..255] range, as well

The "binary string" methods will throw abruptly

The legacy "string" methods (`encodeString` and `decodeString`) will throw asynchronously (in debug mode only, for now, but the default will flip soon), allowing any problems to be caught in logs before changes roll out to production

RELNOTES[NEW]: Added a handful of `goog.crypt` methods with clearer string encoding semantics.

PiperOrigin-RevId: 473318062
Change-Id: I8960fd9517a71d5dd54b5a4d856cab8a688075b8