Add support for console.createTask to goog.Promise.then.
Add helpers to a new asyncstacktag file in debug.

This brings goog.Promise.then roughly in line with what Promise.then stack traces look like, where the stack trace for async links directly to the .then callback site, rather than to the goog.async.run infra. There's still some infra calls in the callstack but it now connects .then() to the callbacks passed into .then(). Currently it links to to the goog.Promise that first started the async callback, which may or may not be related to the .then() callback itself.
RELNOTES: n/a
PiperOrigin-RevId: 492050699
Change-Id: I4942a4dc2547e004c9fd90f9ae282b68065065d7