Disable the Closure Debug Loader by default.
This is being done because pre-bundling code is always a superior strategy and when bundling the debug loader is not needed. Having the loader enabled may hide misconfigurations or log unactionable warnings.  This will eventually make it possible to move the debug loader out of base.js and drop all the associate code from development scenarios.

RELNOTES: Disable the Closure Debug Loader by default.
PiperOrigin-RevId: 513384198
Change-Id: I85571d2781a96c89f1e39466d76ed353d19024c3