Add a constant to Closure KeyCodes API for FireFox +/* key on German layouts and treats it as a character key for FF.
The close bracket key on US keyboard layouts corresponds to
the +/* key on DE keyboard layouts (the same physical key on
the keyboard produces different outputs depending on keyboard layout).
Firefox sends a unique keycode for this (171), which differs from the
keycode sent by Chrome (221). (See:
https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/keyCode)

RELNOTES:Added a keycodes.js constant for the +/* key in Firefox for DE keyboard layouts.
PiperOrigin-RevId: 574474491
Change-Id: Iefdcefa8f4d860c61ffd9e648dc553e871d14fb1